chore: add comparison between AS generated vs plugin generated on the playground app chore: renamed package svgToCompose to svg_to_compose
