+++ playground/.gitignore
@@ -3,9 +3,10 @@
 # Ignore the icon path since this project is intended
 # just for checking if all icons are correctly rendered
 # on Preview tab.
-/app/src/main/kotlin/dev/tonholo/svgToCompose/playground/ui/icon
+/app/src/main/kotlin/dev/tonholo/svg_to_compose/playground/ui/icon
 
 /app/src/main/res/drawable/*
+!/app/src/main/res/drawable/brasil.xml
 !/app/src/main/res/drawable/ic_launcher_background.xml
 !/app/src/main/res/drawable/ic_launcher_foreground.xml
 /app/src/main/res/minimap-*/*
