--- playground/app/src/main/kotlin/dev/tonholo/svgToCompose/playground/MainActivity.kt
@@ -1,18 +1,0 @@
-package dev.tonholo.svgToCompose.playground
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.material3.Text
-import dev.tonholo.svgToCompose.playground.ui.theme.SampleAppTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            SampleAppTheme {
-                Text(text = "This is a sample app. Nothing to show here.")
-            }
-        }
-    }
-}
