+++ playground/app/src/main/kotlin/dev/tonholo/svg_to_compose/playground/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package dev.tonholo.svgToCompose.playground.ui.theme
+package dev.tonholo.svg_to_compose.playground.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -14,13 +14,6 @@ import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
 import androidx.core.view.WindowCompat
-import dev.tonholo.svgToCompose.playground.ui.theme.Pink40
-import dev.tonholo.svgToCompose.playground.ui.theme.Pink80
-import dev.tonholo.svgToCompose.playground.ui.theme.Purple40
-import dev.tonholo.svgToCompose.playground.ui.theme.Purple80
-import dev.tonholo.svgToCompose.playground.ui.theme.PurpleGrey40
-import dev.tonholo.svgToCompose.playground.ui.theme.PurpleGrey80
-import dev.tonholo.svgToCompose.playground.ui.theme.Typography
 
 private val DarkColorScheme = darkColorScheme(
     primary = Purple80,
